requirements.txt
date_detector==1.0.2
date_extractor==5.1.5
datefinder==0.7.1
dateparser==1.1.1
fasttext==0.9.2
gcld3==3.0.13
goose3==3.1.11
htmldate==1.2.1
langcodes==3.3.0
langdetect==1.0.7
lingua==4.15.0
linguars==0.3.2
matplotlib==3.5.1
nltk==3.6.7
numpy==1.22.3
pandas==1.4.1
py3langid==0.2.0
pyfranc==0.1.1
python_dateutil==2.8.2
spacy==3.2.1
textblob==0.17.1
tika==1.24
tqdm==4.63.0
transformers==4.17.0
vaderSentiment==3.3.2
